Convert Audio loader functions to member methods of SongMetadata
